(#86) Support regex in R10K downloaded modules
When selecting roles to test using regexps, onceover expects classes
to exist in `site`, if they are specified in the `Puppetfile` instead,
then they will not be available for scanning _and_ wont be in a directory
that would ordinarily be scanned.

This PR changes onceover behaviour:
1. Process `Puppetfile` first
2. Add `.onceover/etc/puppetlabs/code/environments/production/code/modules`
   to onceover scan path

There is a tight coupling between `TestConfig` and `ControlRepo` classes
that required creating a new class `Deploy` in order to allow R10K to be
called first. Additional debug logging was added to verify correct order
of operation.