(maint) Allow local Gemfile overrides
Previously only gems within the gemspec could be bundled however this prevents
developers using the preferred debug tools etc.  This commit adds the ability
for the Gemfile to include per profile and per project Gemfile additions via the
common ~/.gemfile and Gemfile.local files.  This commit also adds Gemfile.local
to the ignore list, so local files wil not be checked in.