Fix opts[:manifest] setting
Before this patch, running `onceover run spec` with `--manifest PATH` or
when `opts[:manifest]` is set in the `spec/onceover.yaml` file would
result in the Error `#<TypeError: no implicit conversion of nil into
String>`.

This patch fixes that issue by correcting the initialization logic for
configuring the manifests directory, and ensuring the path is used as
the value for `c.manifest` setting in `.onceover/spec/spec_helper.rb`.