Adapted for Facter 4 factsets
The facts function was updated to understand that the top level key
'values' may not exist. A two new factsets were added under spec to
validate things work with both a Facter 4 and pre-Facter 4 factset is
present. Additionally, the list of nodes in the caching control repo was
sorted.

In addition to changes directly related to this issue, fixes required by
the current version of Rubocop were also applied.

Fixes #307