Fix compatibility with rspec-puppet 4+
These settings where used for compatibility with Puppet 7.10 and older
(Puppet 7.10 was released more than 2 years ago).  rspec-puppet 4
removed them:
https://github.com/puppetlabs/rspec-puppet/pull/73

In order to avoid this kind of accident in the future, add an upper
bound on the rspec-puppet dependency.  Version 5.0.0 is also working
with this change, so accept anything before 6.0.0.